requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
supabase==2.3.4
asyncpg==0.29.0

# Background Tasks (using FastAPI built-in)
# No external queue system needed for MVP

# Storage (Cloudflare R2 - S3 compatible)
boto3==1.34.14

# AI Services
httpx==0.26.0
aiohttp==3.9.1

# Face Detection
mediapipe==0.10.15
opencv-python-headless==4.9.0.80
numpy==1.26.3
Pillow==10.2.0

# PDF Generation
weasyprint==60.2
jinja2==3.1.3

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Utilities
python-dotenv==1.0.0
tenacity==8.2.3
structlog==24.1.0

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
